+++ lib/savon/soap/xml.rb
@@ -1,14 +1,7 @@
 require "builder"
 require "gyoku"
 require "rexml/document"
-require "nori"
-
 require "savon/soap"
-
-Nori.configure do |config|
-  config.strip_namespaces = true
-  config.convert_tags_to { |tag| tag.snakecase.to_sym }
-end
 
 module Savon
   module SOAP